Last Version Code

+++ Ti2Discretas/src/model/IDijkstra.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 public interface IDijkstra<T> {
     void addVertex(T vertex);
-    void addEdge(T source, T destination, int weight);
-    Map<T, Integer> dijkstra(T start);
+    void addEdge(T source, T destination, double weight);
+    Map<T, Double> dijkstra(T start);
 
 }
